+++ examples/zpa_cloud_browser_isolation_banner/datasource.tf
@@ -1,0 +1,3 @@
+data "zpa_cloud_browser_isolation_banner" "this" {
+  name = "Default"
+}