+++ zpa/provider_test.go
@@ -56,6 +56,8 @@ func TestMain(m *testing.M) {
 		setupSweeper(resourcetype.ZPAServerGroup, sweepTestServerGroup)
 		setupSweeper(resourcetype.ZPAServiceEdgeGroup, sweepTestServiceEdgeGroup)
 		setupSweeper(resourcetype.ZPAPolicyAccessRule, sweepTestAccessPolicyRuleByType)
+		setupSweeper(resourcetype.ZPACBIBannerController, sweepTestCBIBanner)
+		setupSweeper(resourcetype.ZPACBIExternalIsolationProfile, sweepTestCBIExternalProfile)
 	}
 	resource.TestMain(m)
 }